+++ src/draw_line.h
@@ -9,10 +9,10 @@ namespace cg
 	class LineDrawing : public ClearRenderTarget
 	{
 	public:
-		LineDrawing(unsigned short width, unsigned short height);
+		LineDrawing(unsigned width, unsigned height);
 		virtual ~LineDrawing();
 
-		void DrawLine(unsigned short x_begin, unsigned short y_begin, unsigned short x_end, unsigned short y_end, color color);
+		void DrawLine(unsigned x_begin, unsigned y_begin, unsigned x_end, unsigned y_end, color color);
 
 		void DrawScene();
